added info about Jarvis, updated the overview, changed the short and long term goals
There are a few inconsistencies throughout the project, particularly in the documentation. One bothersome incosistency is 'MkySynth' vs. 'Mkysynth.' I thought we decided upon the latter but my memory might be failing me.

There's going to be a hackathon this weekend and I'm sure we'll clear up these discrepencies and make the set of files fully consistent.

And since I'm using this readme to give my comments on the whole project, I'm going to add that I think the current folder structure needs to be completely reorganized, simply because there's a lot of files hanging around that aren't really a part of the project anymore.

I'm probably going to suggest that most of the 'obsolete' files be moved to an 'archive' folder since some of the code might still prove useful.

I don't know if anyone will ever read this but it's a nice place to write things down, and a habit I'll probably get into (writing longer commentaries in the readme files)