takes raw frames and writes them as binary files
I created this because I'm not sure if the 'raw data' that I'm creating is actually useful. I'm also not sure if x264 is going to be able to parse it. This filter will help with debugging.

It's also a filter that I was going to make at some point anyway.